requirements.txt
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
numpy==1.24.1
Pillow==9.3.0
scikit-learn==1.2.2
scipy==1.10.1

gunicorn==20.1.0
fastapi==0.95.0
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
filelock==3.9.0
h11==0.14.0
idna==3.4
img2vec-pytorch==1.0.1
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
pydantic==1.10.7
python-multipart==0.0.6
requests==2.28.1
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
threadpoolctl==3.1.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.21.1
loguru==0.7.0
python-dotenv==1.0.0
redis==4.5.4
pandas==2.0.0